semiconductor_process_agent/perception/segmentation.py
import numpy as np
from PIL import Image
import torch
import torch.nn as nn
import cv2

# --- Constants for Material IDs ---
MATERIAL_MAP = {
    'vacuum': 0,
    'silicon': 1,
    'silicon_dioxide': 2,
    'polysilicon': 3,
    'photoresist': 4,
    'aluminum': 5,
    'copper': 6,
    'silicon_nitride': 7
}
ID_TO_MATERIAL = {v: k for k, v in MATERIAL_MAP.items()}

# --- Dummy U-Net Model Definition ---
# In a real implementation, this would be a proper U-Net architecture.
# For this example, we'll use a placeholder that simulates the model's behavior.
class DummyUNet(nn.Module):
    def __init__(self):
        super(DummyUNet, self).__init__()
        # This is a placeholder. A real U-Net would have many layers.
        self.conv = nn.Conv2d(3, len(MATERIAL_MAP), kernel_size=1)

    def forward(self, x):
        return self.conv(x)

class SemanticSegmenter:
    """
    Performs semantic segmentation on a schematic image to produce a material map.
    """
    def __init__(self, model_path: str = None):
        """
        Initializes the segmenter and loads a pre-trained model.

        Args:
            model_path (str): Path to the pre-trained PyTorch model file.
        """
        self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
        # In a real scenario, we load a trained model.
        # self.model = torch.load(model_path)
        # For this example, we use a dummy model.
        self.model = DummyUNet().to(self.device)
        self.model.eval()
        print(f"PERCEPTION: Segmenter initialized on device: {self.device}")

    def segment_image(self, image: np.ndarray) -> np.ndarray:
        """
        Segments an image into its constituent materials.

        Args:
            image (np.ndarray): The input image in BGR format (from OpenCV).

        Returns:
            np.ndarray: A 2D array (material map) where each pixel value is a material ID.
        """
        print("PERCEPTION: Starting segmentation...")
        # Preprocess the image for the model
        img_rgb = cv2.cvtColor(image, cv2.COLOR_BGR2RGB)
        img_tensor = torch.from_numpy(img_rgb).permute(2, 0, 1).float() / 255.0
        img_tensor = img_tensor.unsqueeze(0).to(self.device)

        with torch.no_grad():
            output = self.model(img_tensor)
        
        # Post-process the output to get the material map
        # The output is a tensor of shape (1, num_classes, height, width)
        # We take the argmax along the class dimension to get the most likely class for each pixel.
        material_map = torch.argmax(output.squeeze(), dim=0).cpu().numpy().astype(np.uint8)
        
        print("PERCEPTION: Segmentation complete.")
        return material_map

    @staticmethod
    def generate_synthetic_data(width=512, height=256):
        """
        Generates a synthetic schematic and its corresponding segmentation mask.
        This is a simplified example of the data generation process.
        """
        image = Image.new('RGB', (width, height), color='white') # Vacuum
        mask = np.zeros((height, width), dtype=np.uint8) # Vacuum ID is 0

        # Draw Silicon Substrate
        substrate_height = int(height * 0.4)
        image.paste((100, 100, 120), (0, height - substrate_height, width, height))
        mask[height - substrate_height:, :] = MATERIAL_MAP['silicon']

        # Draw Oxide Layer
        oxide_height = 20
        image.paste((150, 200, 220), (0, height - substrate_height - oxide_height, width, height - substrate_height))
        mask[height - substrate_height - oxide_height : height - substrate_height, :] = MATERIAL_MAP['silicon_dioxide']

        return np.array(image), mask
