semiconductor_process_agent/requirements.txt
numpy
opencv-python
scikit-image
torch
torchvision
neo4j
gdspy
smt
scipy
Pillow
